move sorting, filtering and pagination logic to useEmployeeTableLogic composable
